Temp hard code vm display name
Since xrandr is not able to find a display in the tty, it's unable to give a name to the display.
Therefore hardcoded to "Virtual1". Should work with all supported distros. I believe this will only cause issues with Gentoo since they go with "Virtual-1" by default.
Tested only with VirtualBox